Implement /list_completions for Racer
The /list_completions endpoint now returns a list of objects containing
completion text, the full context of the completion text, the file where
it's found, and the cursor location in that file. As in the
/find_definition endpoint, the request takes a list of dirty buffers, a
file path, and cursor location from where the search is performed.
Additionally, the endpoint assumes the Racer engine.

The Context object was refactored to own its buffers and query_file
string. The http requests are now consumed when a context is extracted
from them. There are no copies in this process. All data has move
semantics in this case.

There is a common chunk of code which both find_definition and
list_completions use. I attempted to refactor it into a separate
function, but there were subtle lifetime errors that I couldn't
surmount. This should be resolved in the future.

Both the /find_definition and /list_completions endpoints have identical
request formats. The request object should be refactored into a common
location.