Fix Racer instance sharing
Methods on the ::racer::core::Session object hand out lifetimes tied to
the Session itself. Methods on a struct holding the Session need to have
a concrete &self lifetime which matches the struct's lifetime in order
to use the Session. This becomes problematic when implementing methods
for a trait object that do not contain such a lifetime in the method
prototype. The solution, as seen in librustc/middle/ty/context.rs's
`struct ctxt`, is to store a reference to such an object which has the
lifetime constrained to the lifetime of the holder. In the case of
librustc, the owned object is allocated on the stack, and a reference is
provided for the life of the consumer. In our case, since we can't just
pass something in from the stack, we Box a session and take
responsibility of its allocation. This requires a small bit of unsafe
code. I expect there is a nicer solution that doesn't require the unsafe
code, but it has escaped me for a long time. As they say, "the show must
go on." Suggestions would be very much appreciated.

The query_src parameters passed to racer methods were reinstated since
they are actually used.