Collapsing whitespace in completion context
Without this fix, racerd would return lots of whitespace in the
completion context if the user requested completion for a function like
the following:

fn format( &self )
    -> u32 { }

All the whitespace between `)` and `->` (including the newline) would be
preserved. That's not what the user wants so we now collapse all runs of
whitespace chars to a single space.