Invert build.rs expect() messages.
Result::expect() is oddly named. It takes a panic string as an
argument, and one wants to explain what was expected there so
the code reads like an English sentence. However, the message
is printed on error, when what is expected _isn't_ present:

  thread 'main' panicked at 'OUT_DIR environment variable is set',
  ../src/libcore/option.rs:699

This is confusing as an error message.