Update racer to master.
Racer hasn't been published in a while. I was setting up YouCompleteMe, when I
found that racerd panicked pretty frequently inside racer.

Updated racer solved the problem, but I needed to modify racerd a bit, so this
is the minimal set of modifications that allowed me to build.

Obviously the git dependency path would need to be changed to a proper version
when racer is published, I'm just leaving this here in case this is useful for
someone.